+++ src/main/java/com/telepathicgrunt/commandstructures/ConfiguredFeatureSpawnCommand.java
@@ -1,9 +1,6 @@
 package com.telepathicgrunt.commandstructures;
 
 import com.mojang.brigadier.CommandDispatcher;
-import com.mojang.brigadier.arguments.BoolArgumentType;
-import com.mojang.brigadier.arguments.IntegerArgumentType;
-import com.mojang.brigadier.arguments.LongArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.tree.LiteralCommandNode;
 import net.minecraft.commands.CommandRuntimeException;
@@ -15,34 +12,12 @@ import net.minecraft.commands.arguments.coordinates.Coordinates;
 import net.minecraft.commands.arguments.coordinates.Vec3Argument;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Registry;
-import net.minecraft.data.worldgen.ProcessorLists;
 import net.minecraft.network.chat.TextComponent;
-import net.minecraft.network.protocol.game.ClientboundLevelChunkWithLightPacket;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ChunkHolder;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.level.ChunkPos;
-import net.minecraft.world.level.chunk.LevelChunk;
-import net.minecraft.world.level.levelgen.LegacyRandomSource;
-import net.minecraft.world.level.levelgen.RandomSupport;
-import net.minecraft.world.level.levelgen.WorldgenRandom;
-import net.minecraft.world.level.levelgen.XoroshiroRandomSource;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
-import net.minecraft.world.level.levelgen.feature.configurations.JigsawConfiguration;
-import net.minecraft.world.level.levelgen.feature.structures.JigsawPlacement;
-import net.minecraft.world.level.levelgen.feature.structures.SinglePoolElement;
-import net.minecraft.world.level.levelgen.structure.BoundingBox;
-import net.minecraft.world.level.levelgen.structure.PoolElementStructurePiece;
-import net.minecraft.world.level.levelgen.structure.StructurePiece;
-import net.minecraft.world.level.levelgen.structure.pieces.PieceGenerator;
-import net.minecraft.world.level.levelgen.structure.pieces.PieceGeneratorSupplier;
-import net.minecraft.world.level.levelgen.structure.pieces.StructurePiecesBuilder;
-import net.minecraft.world.level.levelgen.structure.templatesystem.StructureProcessorList;
 
-import java.util.BitSet;
-import java.util.Optional;
 import java.util.Set;
-import java.util.function.Supplier;
 
 public class ConfiguredFeatureSpawnCommand {
     public static void dataGenCommand(CommandDispatcher<CommandSourceStack> dispatcher) {
@@ -80,18 +55,6 @@ public class ConfiguredFeatureSpawnCommand {
 
         cf.place(level, level.getChunkSource().getGenerator(), level.getRandom(), centerPos);
 
-        for(ChunkHolder chunkholder : level.getChunkSource().chunkMap.getChunks()) {
-            LevelChunk levelChunk = chunkholder.getTickingChunk();
-            if(levelChunk != null) {
-                int viewDistance = level.getChunkSource().chunkMap.viewDistance;
-                ClientboundLevelChunkWithLightPacket lightPacket = new ClientboundLevelChunkWithLightPacket(levelChunk, level.getLightEngine(), null, null, true);
-                level.players().forEach(player -> {
-                    int distance = player.chunkPosition().getChessboardDistance(levelChunk.getPos());
-                    if(distance < viewDistance) {
-                        player.trackChunk(levelChunk.getPos(), lightPacket);
-                    }
-                });
-            }
-        }
+        Utilities.refreshChunksOnClients(level);
     }
 }